+++ src/kprc/visitor/statements/DefineFunctionStatementVisitor.cpp
@@ -1,14 +1,14 @@
 #include <memory>
 #include <vector>
 
-#include "../../abstructs/ExprObject.h"
-#include "../../abstructs/FunctionTypeObject.h"
-#include "../../abstructs/StatementObject.h"
 #include "../../FunctionManager.h"
 #include "../../KaprinoAccelerator.h"
 #include "../../StatementVisitor.h"
 #include "../../TypeManager.h"
 #include "../../VariableManager.h"
+#include "../expr/ExprObject.h"
+#include "../FunctionTypeObject.h"
+#include "StatementObject.h"
 #include "KaprinoParserBaseVisitor.h"
 
 class DefineFunctionStatementObject : StatementObject {
