+++ src/kprc/visitor/statements/IfStatementVisitor.cpp
@@ -1,10 +1,10 @@
 #include <memory>
 #include <vector>
 
-#include "../../abstructs/ExprObject.h"
-#include "../../abstructs/StatementObject.h"
 #include "../../StatementVisitor.h"
 #include "../../VariableManager.h"
+#include "../expr/ExprObject.h"
+#include "StatementObject.h"
 #include "KaprinoParserBaseVisitor.h"
 
 class IfStatementObject : StatementObject {
